+++ catalogue/TapeDrivesCatalogueState.hpp
@@ -18,6 +18,7 @@
 #pragma once
 
 #include <limits>
+#include <optional>
 #include <string>
 
 #include "common/dataStructures/DriveStatus.hpp"