+++ catalogue/interfaces/TapeFileCatalogue.hpp
@@ -17,6 +17,7 @@
 
 #pragma once
 
+#include <optional>
 #include <set>
 #include <string>