+++ ReleaseNotes.md
@@ -2,6 +2,7 @@
 
 ### Features
 - cta/CTA#78 - Remove gRPC and ability to change the storage class in EOS from the cta-change-storage-class C++ script
+- cta/CTA#342 - Compile CTA using devtoolset-11
 
 ### Bug Fixes
 - cta/CTA#259 - cta-rmcd should not exit if /dev/sg0 is missing